requirements.txt
python-telegram-bot==20.7
openai
google-generativeai
flask
python-docx
python-dotenv
pandas
openpyxl